feat: support other http agent * add Abstract Class `httpClient` * change ClassName Request => hcpHttpClient * add ConnectionPool.addExternalHttpClient
