+++ example/other-agent/superagent.ts
@@ -1,6 +1,6 @@
 import ConnectionPool from 'http-connection-pool';
-import axios, { AxiosResponse } from "axios";
-import app from './server';
+import superagent, {Response} from "superagent";
+import app from '../server';
 
 const PORT = 3000;
 
@@ -16,9 +16,9 @@ const server = app.listen(PORT);
   const connectionPool = new ConnectionPool(10);  
   for (let i = 0; i < 100; i++) {
     if (i % 10 == 0) {
-      connectionPool.addExternalHttpClient<AxiosResponse>(axios.get, `http://localhost:${PORT}/test`).then(d => console.log(d.data));
+      connectionPool.addExternalHttpClient<Response>(superagent.get, `http://localhost:${PORT}/test`).then(d => console.log(d.text, i));
     }
-    connectionPool.addExternalHttpClient<AxiosResponse>(axios.get, `http://localhost:${PORT}/test`);
+    connectionPool.addExternalHttpClient<Response>(superagent.get, `http://localhost:${PORT}/test`);
   }
   await connectionPool.done();
   server.close();
