+++ package.json
@@ -50,6 +50,7 @@
     "@types/got": "^9.6.12",
     "@types/jest": "^29.5.11",
     "@types/node": "^20.11.5",
+    "@types/superagent": "^8.1.3",
     "axios": "^1.6.7",
     "express": "^4.18.2",
     "got": "^14.1.0",
